can send data to html, but web behavior need fix

+++ src/main/java/com/ticket/tracking/entity/ticket/TicketResponse.java
@@ -8,9 +8,9 @@ public class TicketResponse {
     private String ticketType;
     private String priority;
     private String severity;
-    private int createDate;
-    private int expectedDate;
-    private int resolveDate;
+    private long createDate;
+    private long expectedDate;
+    private long resolveDate;
     private String reporter;
     private String assignee;
 
@@ -109,27 +109,27 @@ public class TicketResponse {
         }
     }
 
-    public int getCreateDate() {
+    public long getCreateDate() {
         return createDate;
     }
 
-    public void setCreateDate(int createDate) {
+    public void setCreateDate(long createDate) {
         this.createDate = createDate;
     }
 
-    public int getExpectedDate() {
+    public long getExpectedDate() {
         return expectedDate;
     }
 
-    public void setExpectedDate(int expectedDate) {
+    public void setExpectedDate(long expectedDate) {
         this.expectedDate = expectedDate;
     }
 
-    public int getResolveDate() {
+    public long getResolveDate() {
         return resolveDate;
     }
 
-    public void setResolveDate(int resolveDate) {
+    public void setResolveDate(long resolveDate) {
         this.resolveDate = resolveDate;
     }
